Re-add custom editor files after incorrect PR
